worker/requirements.txt
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
google-generativeai==0.3.2
langchain==0.1.3
langchain-community==0.0.19
langchain-google-genai==0.0.6
pdfplumber==0.10.3
python-docx==1.0.1
markdown==3.5.2
unstructured==0.11.2
tiktoken==0.5.2
python-dotenv==1.0.0
httpx==0.26.0
pydantic-settings==2.1.0  # Settings management for Pydantic
